requirements.txt
django
djangorestframework
django-filter
mysqlclient
django-debug-toolbar
django-autocomplete-light
djoser
